+++ src/main.rs
@@ -82,7 +82,6 @@ pub(crate) mod http_test_service;
 mod install_id;
 #[allow(dead_code)]
 mod jellyfin;
-#[allow(dead_code)]
 pub(crate) mod lastfm;
 #[allow(dead_code)]
 mod local;